Added Flask functionality
Updated HTML to work with Flask using the default directory.
All references are now based in /static/

Added basic functionality to the guess form. It can now be used to guess and will update to red for an incorrect response and green for a correct response. Both the color and current input are cleared upon new roll.
Added a header for the history list